Migrate to Tauri 2 Stable (#3)
* update npm deps

* update cargo deps

* update compatibility date. need to fix breaking changes.

* begin migration to tauri 2 stable

* more npm dep updates

* updated to use new lib.rs layout

* usage fixes in tauri 2 stable

* fix app no window (may break on Windows!)

* add gen / schemas to gitignore
