+++ package.json
@@ -12,24 +12,33 @@
     "tauri": "tauri"
   },
   "dependencies": {
-    "@iconify-json/tabler": "^1.1.103",
-    "@nuxt/ui": "^2.11.1",
-    "@tauri-apps/api": "2.0.0-alpha.13",
-    "@tauri-apps/plugin-process": "2.0.0-alpha.5",
-    "@tauri-apps/plugin-shell": "2.0.0-alpha.5",
-    "@tauri-apps/plugin-updater": "2.0.0-alpha.5",
-    "@vueuse/integrations": "^10.7.1",
+    "@iconify-json/tabler": "^1.2.17",
+    "@nuxt/ui": "^2.21.1",
+    "@tauri-apps/api": "^2.4.0",
+    "@tauri-apps/plugin-opener": "^2.2.6",
+    "@tauri-apps/plugin-process": "^2.2.0",
+    "@tauri-apps/plugin-shell": "^2.2.0",
+    "@tauri-apps/plugin-updater": "^2.6.1",
+    "@vueuse/integrations": "^10.11.1",
     "lodash-es": "^4.17.21",
-    "sortablejs": "^1"
+    "sortablejs": "^1.15.6"
   },
   "devDependencies": {
     "@nuxt/devtools": "latest",
-    "@tauri-apps/cli": "2.0.0-alpha.20",
+    "@tauri-apps/cli": "^2.4.0",
     "@types/lodash-es": "^4.17.12",
-    "@vueuse/core": "^10.7.1",
-    "@vueuse/nuxt": "^10.7.1",
-    "nuxt": "^3.9.1",
-    "vue": "^3.4.5",
-    "vue-router": "^4.2.5"
+    "@vueuse/core": "^10.11.1",
+    "@vueuse/nuxt": "^10.11.1",
+    "nuxt": "^3.16.1",
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0"
+  },
+  "packageManager": "pnpm@10.6.5+sha1.da220e966c2db3275efb5928cf8f3592681b106b",
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "@parcel/watcher",
+      "esbuild",
+      "vue-demi"
+    ]
   }
 }
